+++ app/Modules/Employees/Views/employees/create.blade.php
@@ -20,7 +20,7 @@
             <div class="card-body">
                 <!-- Employee Number input-->
                 <div class="form-group d-flex align-items-center">
-                    <label class="form-group d-flex align-items-center"
+                    <label class="col-sm-1 text-right text"
                            for="number">{{ trans('fi.employee_number') }}</label>
                     <div class="col-md-4">
                         {!! Form::text('number',old('number'),['id'=>'number', 'class'=>'form-control']) !!}
@@ -28,7 +28,7 @@
                 </div>
                 <!-- First Name input-->
                 <div class="form-group d-flex align-items-center">
-                    <label class="col-md-1 text-right text"
+                    <label class="col-sm-1 text-right text"
                            for="first_name">{{ trans('fi.employee_first_name') }}</label>
                     <div class="col-md-4">
                         {!! Form::text('first_name',old('first_name'),['id'=>'first_name', 'class'=>'form-control']) !!}
@@ -36,7 +36,7 @@
                 </div>
                 <!-- Last Name input-->
                 <div class="form-group d-flex align-items-center">
-                    <label class="col-md-1 text-right text"
+                    <label class="col-sm-1 text-right text"
                            for="last_name">{{ trans('fi.employee_last_name') }}</label>
                     <div class="col-md-4">
                         {!! Form::text('last_name',old('last_name'),['id'=>'last_name', 'class'=>'form-control']) !!}
@@ -44,7 +44,7 @@
                 </div>
                 <!-- Title input-->
                 <div class="form-group d-flex align-items-center">
-                    <label class="col-md-1 text-right text"
+                    <label class="col-sm-1 text-right text"
                            for="title">{{ trans('fi.employee_title') }}</label>
                     <div class="col-md-4">
                         {!! Form::text('title',old('title'),['id'=>'title', 'class'=>'form-control','list'=>'listid']) !!}
@@ -62,7 +62,7 @@
                 </div>
                 <!-- Billing Rate input-->
                 <div class="form-group d-flex align-items-center">
-                    <label class="col-md-1 text-right text"
+                    <label class="col-sm-1 text-right text"
                            for="billing_rate">{{ trans('fi.employee_billing_rate') }}</label>
                     <div class="col-md-4">
                         {!! Form::text('billing_rate',old('billing_rate'),['id'=>'billing_rate', 'class'=>'form-control']) !!}
@@ -70,7 +70,7 @@
                 </div>
                 <!-- Schedule Checkbox-->
                 <div class="form-group d-flex align-items-center">
-                    <label class="col-md-1 text-right text"
+                    <label class="col-sm-1 text-right text"
                            for="schedule">{{ trans('fi.scheduleable') }}</label>
                     <div class="col-md-4">
                         {!! Form::checkbox('schedule',1,old('schedule'),['id'=>'schedule', 'class'=>'checkbox']) !!}
@@ -78,7 +78,7 @@
                 </div>
                 <!-- Active Checkbox-->
                 <div class="form-group d-flex align-items-center">
-                    <label class="col-md-1 text-right text"
+                    <label class="col-sm-1 text-right text"
                            for="active">{{ trans('fi.employee_active') }}</label>
                     <div class="col-md-4">
                         {!! Form::checkbox('active',1,old('active'),['id'=>'active', 'class'=>'checkbox']) !!}
@@ -86,7 +86,7 @@
                 </div>
                 <!-- Driver Checkbox-->
                 <div class="form-group d-flex align-items-center">
-                    <label class="col-md-1 text-right text"
+                    <label class="col-sm-1 text-right text"
                            for="driver">{{ trans('fi.employee_driver') }}</label>
                     <div class="col-md-4">
                         {!! Form::checkbox('driver',1,old('driver'),['id'=>'driver', 'class'=>'checkbox']) !!}