+++ app/Modules/Scheduler/Views/partials/_js_datetimepicker.blade.php
@@ -1,7 +1,17 @@
+<style>
+    .xdsoft_datetimepicker .xdsoft_timepicker {
+        width: 75px;
+        float: left;
+        text-align: center;
+        margin-left: 8px;
+        margin-top: 0;
+    }
+</style>
 <script>
     $(document).ready(function () {
         $(".from").datetimepicker({
             format: 'Y-m-d H:i',
+            formatTime: '{{ config('fi.use24HourTimeFormat') ? 'H:i' : 'g:i A' }}',
             defaultDate: new Date(),
             defaultTime: '08:00',
             step: {!! config('fi.schedulerTimestep') !!},
@@ -11,6 +21,7 @@
         });
         $(".to").datetimepicker({
             format: 'Y-m-d H:i',
+            formatTime: '{{ config('fi.use24HourTimeFormat') ? 'H:i' : 'g:i A' }}',
             defaultDate: new Date(),
             defaultTime: '16:00',
             step: {!! config('fi.schedulerTimestep') !!},
@@ -20,6 +31,7 @@
         });
         $(".until").datetimepicker({
             format: 'Y-m-d H:i',
+            formatTime: '{{ config('fi.use24HourTimeFormat') ? 'H:i' : 'g:i A' }}',
             defaultDate: '+1970/02/01',//+1 month
             defaultTime: '16:00',
             step: {!! config('fi.schedulerTimestep') !!},