+++ app/Modules/Scheduler/Views/schedule/tableRecurringEvent.blade.php
@@ -8,11 +8,7 @@
 
             <a href="javascript:void(0)" class="btn btn-secondary bulk-actions" id="btn-bulk-trash"><i
                         class="fa fa-trash"></i> {{ trans('fi.bulk_event_trash') }}</a>
-            {{--<div class="btn-group">--}}
-            {{--{!! Form::open(['method' => 'GET', 'id' => 'filter', 'class'=>"form-inline"]) !!}--}}
-            {{--{!! Form::select('company_profile', $companyProfiles, request('company_profile'), ['class' => 'workorder_filter_options form-control ']) !!}--}}
-            {{--{!! Form::close() !!}--}}
-            {{--</div>--}}
+
             <a href="{!! route('scheduler.createrecurringevent') !!}" class="btn btn-primary "><i
                         class="fa fa-fw fa-plus"></i> {{ trans('fi.create_recurring_event') }}</a>
         </div>
@@ -36,5 +32,4 @@
 @stop
 @section('javascript')
     @include('partials._js_bulk_ajax')
-    {{--@include('partials._js_datatables')--}}
 @stop