+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - documentation
 - "enabled module" in sidebar (system setting)
 - enabled update checker
+- added JQuery-UI theme config
+- clean and optimize scheduler
 
 ## 4.0.2
 - added resource quantity selection to createworkoorder modal